Add races
