+++ simulation/common/__init__.py
@@ -4,6 +4,9 @@ from simulation.common.helpers import timeit
 from simulation.common.producer import Producer
 from simulation.common.storage import Storage
 from simulation.common.car import Car
+from simulation.common.race import Race
 
 DayBreak = Car('DayBreak')
 BrightSide = Car('BrightSide')
+ASC = Race('ASC')
+FSGP = Race('FSGP')
