+++ simulation/common/car.py
@@ -1,4 +1,5 @@
 from simulation.config import config_directory
+from simulation.utils import Singleton
 import os
 import json
 
@@ -7,17 +8,7 @@ This Singleton stores constants for UBC Solar's vehicles.
 """
 
 
-class CarMetaclass(type):
-    _instances = {}
-
-    def __call__(cls, *args, **kwargs):
-        if cls not in cls._instances:
-            instance = super().__call__(*args, **kwargs)
-            cls._instances[cls] = instance
-        return cls._instances[cls]
-
-
-class Car(metaclass=CarMetaclass):
+class Car(metaclass=Singleton):
     def __init__(self, name: str):
         config_path = os.path.join(config_directory, f"{name}.json")
