+++ simulation/config/ASC.json
@@ -1,0 +1,6 @@
+{
+  "charging_begin" : 7,
+  "charging_end": 20,
+  "driving_begin": 9,
+  "driving_end": 18
+}
