+++ simulation/utils/__init__.py
@@ -1,3 +1,4 @@
 from simulation.utils.InputBounds import InputBounds
 from simulation.utils.SimulationBuilder import SimulationBuilder
 from simulation.utils.Plotting import Plotting, GraphPage
+from simulation.utils.Singleton import Singleton
